backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
pillow
